changed frametime to be uint64_t from realsense api when frame arrives, working on fixes for setting dmax and dmin in shaders

+++ include/interface.h
@@ -97,7 +97,7 @@ public:
 
 	void frames(unsigned char * colorArray, float * bigDepthArray);
 
-	bool frames(double &frameTime, unsigned char * colorArray, std::vector<uint16_t> &depthArray, float * infraredArray, float * bigDepthArray, int * colorDepthMapping);
+	bool frames(uint64_t &frameTime, unsigned char * colorArray, std::vector<uint16_t> &depthArray, float * infraredArray, float * bigDepthArray, int * colorDepthMapping);
 
 	//void frames(cv::Mat &color, cv::Mat &depth, cv::Mat &infra, float & fullColor);
 
@@ -206,7 +206,7 @@ private:
 	float *m_rawColor;
 	float *m_rawBigDepth;
 
-	double m_frameArrivalTime = 0;
+	uint64_t m_frameArrivalTime = 0;
 
 	int m_frame_width;
 	int m_frame_height;
